Rename: 클라이언트, 서버 폴더 분리
