+++ client/js/createVillagePosition.js
