+++ package.json
@@ -1,0 +1,22 @@
+{
+  "name": "fe-postbox",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/youryu0212/fe-postbox.git"
+  },
+  "author": "geon-hoi",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/youryu0212/fe-postbox/issues"
+  },
+  "homepage": "https://github.com/youryu0212/fe-postbox#readme",
+  "dependencies": {
+    "express": "^4.17.3"
+  }
+}
